refactor nanming of evaluations
