+++ src/view/pages/statement/components/evaluations/components/evaluation/simpleEvaluation/SimpleEvaluation.scss
