+++ src/view/pages/statement/components/evaluations/components/suggestionCards/suggestionCard/SuggestionCard.scss
